requirements.txt
#Hardware and MPD Connection
luma.oled==3.8.1
luma-lcd
RPi.GPIO==0.7.0
Pillow==9.5.0
